requirements/requirements-fastapi.txt
fastapi==0.115.11
uvicorn==0.34.0
SQLAlchemy==2.0.39
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
redis==4.5.5
PyMuPDF==1.25.5
pymupdf4llm==0.0.20
arxiv==2.2.0
langchain==0.3.21
langchain-core==0.3.46
langchain-community==0.3.20
langchain-text-splitters==0.3.7
huggingface-hub==0.26.5
transformers==4.47.0
sentence-transformers==3.4.1
FlagEmbedding==1.3.4
chroma-hnswlib==0.7.6
chromadb==0.6.3
celery==5.5.1